fix: address eleventh round of Copilot review comments on PR #48
- segformer_daemon/export_segformer_onnx: drop redundant .astype(np.float32)
  before passing to preprocess_bands() — avoids a full-frame copy since
  preprocess_bands() already coerces to float32 on entry
- segformer_daemon: catch PermissionError from makedirs and log an actionable
  message suggesting --socket /tmp/... for non-root non-systemd runs
- ticktalk_main: bytearray + extend() for daemon response accumulation,
  matching the same fix already applied to handle_connection()

+++ tools/export_segformer_onnx.py
@@ -206,7 +206,7 @@ class _CalibrationReader:
             try:
                 import rasterio
                 with rasterio.open(path) as src:
-                    img = src.read().astype(np.float32)
+                    img = src.read()  # preprocess_bands handles float32 conversion
                 if img.shape[0] != self._n_bands:
                     print(f"  Skipping {path}: expected {self._n_bands} bands, got {img.shape[0]}")
                     continue
